pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vhir-cli"
version = "0.5.3"
description = "ValiHuntIR CLI: human-in-the-loop approval, evidence management, forensic command execution"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AppliedIncidentResponse.com"},
]
keywords = ["forensics", "incident-response", "cli", "approval-workflow", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "pyyaml>=6.0",
    "argcomplete>=3.0",
]

[project.scripts]
vhir = "vhir_cli.main:main"

[project.urls]
Homepage = "https://github.com/AppliedIR/valihuntir"
Repository = "https://github.com/AppliedIR/valihuntir"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vhir_cli"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (let formatter handle)
    "SIM105", # contextlib.suppress — verbose for existing code
    "SIM108", # ternary operator — readability preference
    "SIM115", # context manager for open — not always applicable
    "SIM117", # combine with statements — verbose for existing code
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/vhir_cli"]
branch = true
